ruamel_yaml: Change to source directory before running `setup.py`.
We found some issues occurred on Windows when not switch to the source
directory before running `setup.py`, but instead using the path to the
source directory when running from the recipe directory. Surprisingly
this issue did not crop up on Linux or Mac. To solve this issue, we
switch to the source directory first and then run `setup.py` to make
sure everything is relative to the right directory. As we are making
this directory change in the subshell only, we remain in the recipe
directory for the rest of the script.